feat: add vault open interest chart
Adds the shared open-interest doughnut chart to single-vault overview pages for borrowable EVaults, and reuses that block from the borrow pair overview.
